backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-bcrypt
pyjwt
flask-limiter
flask-talisman
email-validator
python-dotenv
